feat(scientific-projects): add admin edit page for scientific projects
